Fix PostgreSQL feedback data types

+++ backend/models/Feedback.js
@@ -5,13 +5,20 @@ const Feedback = sequelize.define('Feedback', {
   id: { type: DataTypes.UUID, defaultValue: DataTypes.UUIDV4, primaryKey: true },
   appointment_id: { type: DataTypes.UUID, allowNull: false, unique: true },
   client_id: { type: DataTypes.UUID, allowNull: false },
-  rating: { type: DataTypes.TINYINT, allowNull: false },
+
+  rating: { type: DataTypes.SMALLINT, allowNull: false },
   comment: { type: DataTypes.TEXT, allowNull: true },
-  booking_experience: { type: DataTypes.TINYINT, allowNull: true },
-  staff_service: { type: DataTypes.TINYINT, allowNull: true },
-  studio_experience: { type: DataTypes.TINYINT, allowNull: true },
-  cleanliness: { type: DataTypes.TINYINT, allowNull: true },
-  overall_satisfaction: { type: DataTypes.TINYINT, allowNull: true },
-}, { tableName: 'feedback', underscored: true, updatedAt: false });
+
+  booking_experience: { type: DataTypes.SMALLINT, allowNull: true },
+  staff_service: { type: DataTypes.SMALLINT, allowNull: true },
+  studio_experience: { type: DataTypes.SMALLINT, allowNull: true },
+  cleanliness: { type: DataTypes.SMALLINT, allowNull: true },
+  overall_satisfaction: { type: DataTypes.SMALLINT, allowNull: true },
+
+}, {
+  tableName: 'feedback',
+  underscored: true,
+  updatedAt: false
+});
 
 module.exports = Feedback;
